rename utils.js to i18n.js
Add support like 'pl-PL' to 'pl'
