--- _src/src/utils.js
@@ -1,9 +1,0 @@
-export default {
-    getLanguage() {
-        // Chrome and Firefox 32+
-        return navigator.languages || [
-          navigator.language ||
-          navigator.browserLanguage
-        ]
-    }
-}
